+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 1. [#49](https://github.com/influxdata/influxdb-client-php/pull/49): Added support for Guzzle 7
 1. [#51](https://github.com/influxdata/influxdb-client-php/pull/51): CSV parser is able to parse export from UI
 
+### CI
+1. [#54](https://github.com/influxdata/influxdb-client-php/pull/54): Updated default docker image to v2.0.2
+ 
 ### Documentation
-
 1. [#46](https://github.com/influxdata/influxdb-client-php/pull/46): Adding info about how to get the query in JSON format
 
 ## 1.8.0 [2020-10-30]